taller_back/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6

# ORM / DB
SQLAlchemy==2.0.36
pymysql==1.1.1

# Seguridad / JWT
python-jose[cryptography]==3.3.0

# Hash de contraseñas
passlib==1.7.4
bcrypt==4.2.0
argon2-cffi==23.1.0

# Validaciones / config
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
email-validator==2.2.0
python-multipart==0.0.9